check for user before data

+++ controllers/BooksController.php
@@ -23,44 +23,36 @@ class BooksController {
     }
 
     /**
-     * Check if a user is authenticated 
-     * If not, it returns an Unauthorized message
-     */
-
-    public function userAuthenticated(){
-        if(!$this->user || $this->user && !$this->user['UserID']){
-            echo json_encode(["message" => "Unauthorized"]);
-        }
-    }
-
-    /**
      * Check if a book exists given it's id. 
      * If not it retuns false.
      */
 
     public function findBookWithId(int $bookid){
-        $book = $this->dbQuery->select(["*"])
-                    ->from('books')
-                    ->where(["id" => $bookid])
-                    ->fetchOne();
-
-        if (count($book) === 0) {
-            echo json_encode(["message" => "Book not found."]);
+        try{
+            if($this->user && $this->user['UserID']){
+                $book = $this->dbQuery->select(["*"])->from('books')->where(["id" => $bookid])->fetchOne();
+    
+                if (count($book) === 0) {
+                    return json_encode(["message" => "Book not found."]);
+                }
+                
+                return json_encode($book);
+            }else{
+                echo json_encode(["message" => "Unauthorized"]);
+            }
+        }catch (PDOException $e) {
+            echo json_encode(["message" => "Error: " . $e]);
         }
-        
-        return json_encode($book);
     }
 
     public function getReadingYears(){
-        $this->userAuthenticated();
-
         try {
-            $years = $this->dbQuery->select(["YEAR(readed) as year", "COUNT(YEAR(readed)) as yearCount"])
-                        ->from('books')
-                        ->groupby("YEAR(readed)")
-                        ->fetchAll();
-
-            echo json_encode($years);
+            if($this->user && $this->user['UserID']){
+                $years = $this->dbQuery->select(["YEAR(readed) as year", "COUNT(YEAR(readed)) as yearCount"])->from('books')->groupby("YEAR(readed)")->fetchAll();
+                echo json_encode($years);
+            }else{
+                echo json_encode(["message" => "Unauthorized"]);
+            }
         } catch (PDOException $e) {
             echo json_encode(["message" => "Error: " . $e]);
         }
@@ -71,14 +63,13 @@ class BooksController {
      */
 
     public function getAllBooks() {
-        $this->userAuthenticated();
-
         try {
-            $books = $this->dbQuery->select(["*"])
-                        ->from('books')
-                        ->fetchAll();
-
-            echo json_encode($books);
+            if($this->user && $this->user['UserID']){
+                $books = $this->dbQuery->select(["*"])->from('books')->fetchAll();
+                echo json_encode($books);
+            }else{
+                echo json_encode(["message" => "Unauthorized"]);
+            }
         } catch (PDOException $e) {
             echo json_encode(["message" => "Error: " . $e]);
         }
@@ -89,15 +80,13 @@ class BooksController {
      */
 
      public function getBooksByYear(string $year) {
-        $this->userAuthenticated();
-
         try {
-            $books = $this->dbQuery->select(["*"])
-                        ->from('books')
-                        ->where(['YEAR(readed)' => $year])
-                        ->fetchAll();
-
-            echo json_encode($books);
+            if($this->user && $this->user['UserID']){
+                $books = $this->dbQuery->select(["*"])->from('books')->where(['YEAR(readed)' => $year])->fetchAll();
+                echo json_encode($books);
+            }else{
+                echo json_encode(["message" => "Unauthorized"]);
+            }
         } catch (PDOException $e) {
             echo json_encode(["message" => "Error: " . $e]);
         }
@@ -108,14 +97,13 @@ class BooksController {
     */
 
     public function insertBook(string $name, string $author, string $genre, string $readed, int $rating, int $en) {
-        $this->userAuthenticated();
-
         try {
-            $book = $this->dbQuery->insertInto('books')
-                        ->data(["name" => $name, "author" => $author, "genre" => $genre, "readed" => $readed, "rating" => $rating, "en" => $en])
-                        ->execute();
-
-            echo json_encode(["message" => "Boek " . $name . " is toegevoegd"]);
+            if($this->user && $this->user['UserID']){
+                $book = $this->dbQuery->insertInto('books')->data(["name" => $name, "author" => $author, "genre" => $genre, "readed" => $readed, "rating" => $rating, "en" => $en])->execute();
+                echo json_encode(["message" => "Boek " . $name . " is toegevoegd"]);
+            }else{
+                echo json_encode(["message" => "Unauthorized"]);
+            }
         } catch (PDOException $e) {
             echo json_encode(["message" => "Error: " . $e]);
         }
@@ -127,18 +115,16 @@ class BooksController {
     */
 
     public function updateBook(int $bookId, string $name, string $author, string $genre, string $readed, int $rating, int $en) {
-        $this->userAuthenticated();
-        
         try {
-            $book = $this->findBookWithId($bookId);
+            if($this->user && $this->user['UserID']){
+                $book = $this->findBookWithId($bookId);
 
-            if($book->id){
-                $book = $this->dbQuery->update('books')
-                        ->set(["name" => $name, "author" => $author, "genre" => $genre, "readed" => $readed, "rating" => $rating, "en" => $en])
-                        ->where(["id" => $bookId])
-                        ->execute();
-
-                echo json_encode(["success" => true, "message" => "Book '$name' has been updated."]);
+                if($book->id){
+                    $book = $this->dbQuery->update('books')->set(["name" => $name, "author" => $author, "genre" => $genre, "readed" => $readed, "rating" => $rating, "en" => $en])->where(["id" => $bookId])->execute();
+                    echo json_encode(["success" => true, "message" => "Book '$name' has been updated."]);
+                }
+            }else{
+                echo json_encode(["message" => "Unauthorized"]);
             }
         } catch (PDOException $e) {
             echo json_encode(["message" => "Error: " . $e]);
@@ -151,19 +137,21 @@ class BooksController {
     */
 
     public function deleteBook(int $bookId, string $bookName) {
-        $this->userAuthenticated();
-        
         try {
-            $book = $this->findBookWithId($bookId);
-            $book = json_decode($book);
-            
-            if($book->id){
-                $book = $this->dbQuery->delete()
-                        ->from("books")
-                        ->where(["id" => $bookId])
-                        ->execute();
+            if($this->user && $this->user['UserID']){
+                $book = $this->findBookWithId($bookId);
+                $book = json_decode($book);
+                
+                if($book->id){
+                    $book = $this->dbQuery->delete()
+                            ->from("books")
+                            ->where(["id" => $bookId])
+                            ->execute();
 
-                echo json_encode(["message" => "Book with ID $bookName has been deleted."]);
+                    echo json_encode(["message" => "Book with ID $bookName has been deleted."]);
+                }
+            }else{
+                echo json_encode(["message" => "Unauthorized"]);
             }
         } catch (PDOException $e) {
             echo json_encode(["message" => "Error: " . $e]);
